# references/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/topics/{id}": {
      "get": {
        "operationId": "Discussions_getTopic",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Discussions"
        ]
      },
      "patch": {
        "operationId": "Discussions_patchTopic",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchTopicDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Discussions"
        ]
      }
    },
...
    "/api/v1/admin/topics/{id}": {
      "patch": {
        "operationId": "AdminDiscussions_pin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinTopicDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "AdminDiscussions"
        ]
      },
      "delete": {
        "operationId": "AdminDiscussions_removeTopic",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "AdminDiscussions"
        ]
      }
    },
    "/api/v1/admin/messages/{id}/hide": {
      "post": {
        "operationId": "AdminDiscussions_hide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "AdminDiscussions"
        ]
      }
    },
    "/api/v1/admin/messages/{id}/unhide": {
      "post": {
        "operationId": "AdminDiscussions_unhide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "AdminDiscussions"
        ]
      }
    },
...
    "/api/v1/admin/trash": {
      "get": {
        "operationId": "AdminTrash_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "page",
                "form",
                "topic",
                "topic_message",
                "chat_message",
                "group",
                "user"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminTrash"
        ]
      }
    },
    "/api/v1/admin/trash/{type}/{id}/restore": {
      "post": {
        "operationId": "AdminTrash_restore",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Type inconnu, élément absent ou non supprimé → `404`.",
        "tags": [
          "AdminTrash"
        ]
      }
    },
...
  },
  "info": {
    "title": "Strategos",
    "description": "API de Strategos (`/api/v1`). Session par cookie `strategos_session` ; toute requête qui modifie des données envoie le jeton de `GET /auth/csrf` dans `X-CSRF-Token`. Erreurs : `{ code, message, details }`. Référence : references/conception/13-api.md.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "strategos_session"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
    "schemas": {
...
      "PatchTopicDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "closed": {
            "type": "boolean"
          }
        }
      },
...
      "PinTopicDto": {
        "type": "object",
        "properties": {
          "pinned": {
            "type": "boolean"
          }
        },
        "required": [
          "pinned"
        ]
      },
...
    }
  }
}